% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregime-package.R
\docType{package}
\name{ecoregime-package}
\alias{ecoregime}
\alias{ecoregime-package}
\title{ecoregime: Analysis of Ecological Dynamic Regimes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for implementing the Ecological Dynamic Regime framework, including functions to characterize and compare groups of ecological trajectories (Sánchez-Pinillos et al., 2023 \doi{10.1002/ecm.1589}); assess the ecological resilience of a disturbed system using a reference dynamic regime (Sánchez-Pinillos et al., 2024 \doi{10.1016/j.biocon.2023.110409}); and forecast ecological trajectories from a dynamic regime (Sánchez-Pinillos et al. 2026, \doi{10.1111/2041-210x.70372}). Additional functions are also available for visualizing ecological dynamic regimes, their representative trajectories, as well as predicted trajectories in a multidimensional state space.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mspinillos.github.io/ecoregime/}
  \item \url{https://github.com/MSPinillos/ecoregime}
  \item Report bugs at \url{https://github.com/MSPinillos/ecoregime/issues}
}

}
\author{
\strong{Maintainer}: Martina Sánchez-Pinillos \email{martina.sanchez.pinillos@gmail.com} (\href{https://orcid.org/0000-0002-1499-4507}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Martina Sánchez-Pinillos \email{martina.sanchez.pinillos@gmail.com} (\href{https://orcid.org/0000-0002-1499-4507}{ORCID}) [copyright holder]
}

}
\keyword{internal}
