% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregime-package.R
\docType{package}
\name{ecoregime-package}
\alias{ecoregime}
\alias{ecoregime-package}
\title{ecoregime: Analysis of Ecological Dynamic Regimes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for implementing the Ecological Dynamic Regime framework (Sánchez-Pinillos et al., 2023 \doi{10.1002/ecm.1589}) to characterize and compare groups of ecological trajectories in multidimensional spaces defined by state variables. The package includes the RETRA-EDR algorithm to identify representative trajectories, functions to generate, summarize, and visualize representative trajectories, and several metrics to quantify the distribution and heterogeneity of trajectories in an ecological dynamic regime and quantify the dissimilarity between two or more ecological dynamic regimes. The package also includes a set of functions to assess ecological resilience based on ecological dynamic regimes (Sánchez-Pinillos et al., 2024 \doi{10.1016/j.biocon.2023.110409}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mspinillos.github.io/ecoregime/}
  \item \url{https://github.com/MSPinillos/ecoregime}
  \item Report bugs at \url{https://github.com/MSPinillos/ecoregime/issues}
}

}
\author{
\strong{Maintainer}: Martina Sánchez-Pinillos \email{martina.sanchez.pinillos@gmail.com} (\href{https://orcid.org/0000-0002-1499-4507}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Martina Sánchez-Pinillos \email{martina.sanchez.pinillos@gmail.com} (\href{https://orcid.org/0000-0002-1499-4507}{ORCID}) [copyright holder]
}

}
\keyword{internal}
