% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregime-package.R
\docType{package}
\name{ecoregime-package}
\alias{ecoregime}
\alias{ecoregime-package}
\title{ecoregime: What the Package Does (One Line, Title Case)}
\description{
Implementation of the EDR framework (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSPinillos/ecoregime}
  \item Report bugs at \url{https://github.com/MSPinillos/ecoregime/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
