% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregime-package.R
\docType{package}
\name{ecoregime-package}
\alias{ecoregime}
\alias{ecoregime-package}
\title{ecoregime: Analysis of Ecological Dynamic Regimes}
\description{
A toolbox for implementating the EDR framework (Sánchez-Pinillos et al.; DOI) to characterize and compare groups of ecological trajectories in multidimensional spaces defined by state variables. The package includes the RETRA-EDR algorithm to identify representative trajectories, functions to generate, summarize, and visualize representative trajectories, and several m etrics to quantify the variability and heterogeneity of trajectories in an ecological dynamic regime and quantify the dissimimilarity between two or more ecological dynamic regimes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSPinillos/ecoregime}
  \item Report bugs at \url{https://github.com/MSPinillos/ecoregime/issues}
}

}
\author{
\strong{Maintainer}: Martina Sánchez-Pinillos \email{martina.sanchez.pinillos@gmail.com} (\href{https://orcid.org/0000-0002-1499-4507}{ORCID})

}
\keyword{internal}
