% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPD.R
\name{MPD}
\alias{MPD}
\title{Mean predicted deviation (MPD)}
\usage{
MPD(x, pc_predicted = 1)
}
\arguments{
\item{x}{Object of class \code{PETRA} returned by \code{petra_edr()} using \code{return_args = T}.}

\item{pc_predicted}{Numeric value in the range 0-1, indicating the percentage
of predicted states used to compute MPD.}
}
\value{
\code{MPD()} returns a numeric value quantifying the average dissimilarity between
the target used to compute \code{x} and the trajectories predicted from the precedent
and following states included in \code{x}. Note that the larger the value of MPD, the
lower the accuracy of \code{x}.
}
\description{
Metric to quantify the accuracy of predicted trajectories using PETRA-EDR.
}
\details{
MDP estimates the accuracy of predicted trajectories (\code{x}) returned by \code{petra_edr()}.
MDP is based on the assumption that only one trajectory passes by each target,
and therefore, subsequent trajectories predicted from any point of \code{x} should
contain the original states in the target.

\eqn{
MPD = \frac{\sum_{i=1}^{n}\sum_{j=1}^{m}d(x_i, \hat{Z}_j)}{n m}
}

where \eqn{d(x_i, \hat{Z}_i)} is the dissimilarity between the initial or final
state \eqn{x_i} of the target and the predicted trajectory \eqn{\hat{Z}_j},
defined as the minimum dissimilarity between \eqn{x_i} and the predicted states
of \eqn{\hat{Z}_j}; \eqn{n} is the number of states of the target; and \eqn{m}
is the number of states of the trajectory predicted from the target (\eqn{\hat{Z}_i}).

\code{pc_predicted} values lower than 1 may reduce the computation time.
}
\examples{
# Define state_var including the state variables for the states in the EDR and
# the target
EDR_var <- EDR_data$EDR1$abundance
target_var <- data.frame(sp1 = 30, sp2 = 10, sp3 = 13, sp4 = 12, sp5 = 5, sp6 = 2,
                         sp7 = 6, sp8 = 3, sp9 = 7, sp10 = 8, sp11 = 2, sp12 = 3)
state_var <- data.frame(rbind(EDR_var[, -c("EDR", "traj", "state")], target_var))

# Compute PETRA-EDR
petra <- petra_edr(state_var = state_var,
                   trajectories = c(EDR_var$traj, "target"),
                   states = as.integer(c(EDR_var$state, 0)),
                   targets = "target",
                   k = 5L,
                   minPts = 2L,
                   d_function = "vegan::vegdist",
                   d_args = list(x = state_var, method = "bray"),
                   return_args = TRUE,
                   direction = 2)

# Calculate MPD
MPD(x = petra)

}
\references{
Sánchez-Pinillos M., Fortin, M-J., Messier, C., Kneeshaw, D. 2026.
Forecasting ecological trajectories from ecological dynamic regimes to improve
resilience analysis. \emph{Methods in Ecology and Evolution} (in press).
}
\seealso{
\code{\link[=petra_edr]{petra_edr()}} for predicting ecological trajectories from EDRs.

\code{\link[=plot.PETRA]{plot.PETRA()}} for plotting predicted trajectories in an ordination space
representing the state space of the EDR.
}
\author{
Martina Sánchez-Pinillos
}
