% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.RETRA}
\alias{summary.RETRA}
\title{Summarize representative trajectories}
\usage{
\method{summary}{RETRA}(object, ...)
}
\arguments{
\item{object}{Object of class \code{RETRA} returned from function \code{\link[=retra_edr]{retra_edr()}}.}

\item{...}{(not used)}
}
\value{
Data frame with nine columns and one row for each representative trajectory
in \code{object}. The columns in the returned data frame contain the following
information:
\itemize{
\item \code{ID}: Identifier of the representative trajectories returned from \code{\link[=retra_edr]{retra_edr()}}.
\item \code{Size}: Number of states forming each representative trajectory.
\item \code{Length}: Sum of the dissimilarities in \code{d} between every pair of consecutive
states forming the representative trajectories.
\item \code{Avg_link}: Mean value of the dissimilarities between every pair of states
linking two representative segments in the representative trajectories.
\item \code{Sum_link}: Sum of the dissimilarities between every pair of states linking
two representative segments in the representative trajectories.
\item \code{Avg_density}: Mean value of the number of segments represented by each real
segment of the representative trajectory (i.e., excluding links).
\item \code{Max_density}: Maximum number of segments represented by at least of the
segments of the representative trajectory (excluding links).
\item \code{Avg_depth}: Mean value of the kd-tree depths, that is, the number of partitions
of the ordination space until finding a region with \code{minSegs} segments or less.
\item \code{Max_depth}: Maximum depth in the kd-tree, that is, the number of partitions
of the ordination space until finding a region with \code{minSegs} segments or less.
}
}
\description{
Summary of the properties of each representative trajectory obtained
with RETRA-EDR.
}
\note{
Note that the \code{density} of the dense regions identified in the EDR tend to \code{minSegs}.
In contrast, greater values of \code{kdTree_depth} indicate a higher stability of
dense regions over sequential partitions of the ordination space.
}
\examples{
d = EDR_data$EDR1$state_dissim
trajectories = EDR_data$EDR1$abundance$traj
states = EDR_data$EDR1$abundance$state
RT <- retra_edr(d = d, trajectories = trajectories, states = states, minSegs = 5)
summary(RT)

}
\seealso{
\code{\link[=retra_edr]{retra_edr()}} for identifying representative trajectories in EDRs.
}
